[WON-72] fix: Medium 이슈 일괄 수정 (#78)
- WON-73: HARRequest.BodySize 타입을 int에서 int64로 통일
- WON-75: WebSocket 프레임 인터셉션 미지원 문서화
- WON-76: Breakpoint onPause nil 시 panic으로 검증
- WON-78: applyHeaderMod에 default case 추가

+++ internal/session/export.go
@@ -42,7 +42,7 @@ type HARRequest struct {
 	Headers     []HARNameValue  `json:"headers"`
 	QueryString []HARNameValue  `json:"queryString"`
 	HeadersSize int             `json:"headersSize"`
-	BodySize    int             `json:"bodySize"`
+	BodySize    int64           `json:"bodySize"`
 }
 
 // HARResponse represents an HTTP response in HAR format.
